requirements.txt
asgiref==3.9.1
Django==5.2.6
django-ckeditor==6.7.3
django-js-asset==3.1.2
sqlparse==0.5.3
whitenoise==6.10.0
